Refactor tsf into base class and methods
